improved waveform precision

+++ src/mathematics.cpp
@@ -2,21 +2,21 @@
 
 #include <cmath>
 
-const vector<float> Mathematics::calculateDistributedPositions(const float start, const float range, const int count)
+const vector<double> Mathematics::calculateDistributedPositions(const double start, const double range, const int count)
 {
-	const float step = range / static_cast<float>(count + 1);
+	const double step = range / static_cast<double>(count + 1);
 
-	vector<float> positions = {};
+	vector<double> positions = {};
 
 	for(int index = 0; index < count; index++)
 	{
-		positions.push_back(start + step * static_cast<float>(index + 1));
+		positions.push_back(start + step * static_cast<double>(index + 1));
 	}
 
 	return positions;
 }
 
-const fmat33 Mathematics::createTranslationMatrix(const float x, const float y)
+const fmat33 Mathematics::createTranslationMatrix(const double x, const double y)
 {
 	fmat33 result = fmat33(1.0f);
 
@@ -26,7 +26,7 @@ const fmat33 Mathematics::createTranslationMatrix(const float x, const float y)
 	return result;
 }
 
-const fmat33 Mathematics::createRotationMatrix(const float angle)
+const fmat33 Mathematics::createRotationMatrix(const double angle)
 {
 	fmat33 result = fmat33(1.0f);
 
@@ -38,7 +38,7 @@ const fmat33 Mathematics::createRotationMatrix(const float angle)
 	return result;
 }
 
-const fmat33 Mathematics::createScalingMatrix(const float x, const float y)
+const fmat33 Mathematics::createScalingMatrix(const double x, const double y)
 {
 	fmat33 result = fmat33(1.0f);
